requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-socketio==5.5.1
eventlet==0.39.1
groq==0.25.0
requests==2.32.3
python-dotenv==1.1.0
duckduckgo-search==7.5.3
